% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_orgs.R
\name{cfg_orgs}
\alias{cfg_orgs}
\alias{lookupOrgGuid}
\alias{orgAuth}
\alias{orgScope}
\alias{orgInfo}
\alias{org_info}
\alias{orgIntegrations}
\alias{org_integrations}
\alias{orgShortName}
\alias{orgTimeZone}
\alias{get_store_id}
\alias{get_store_id_from_db}
\alias{get_org_stores}
\alias{orgIndexFull}
\title{org functions}
\usage{
lookupOrgGuid(name)

orgAuth(org, .api)

orgScope(org, .app)

orgInfo(org, extra_cols = FALSE)

org_info(org, extra_cols = FALSE)

orgIntegrations(org)

org_integrations(org)

orgShortName(org = NULL)

orgTimeZone(org = NULL)

get_store_id(org, store)

get_store_id_from_db(org, store)

get_org_stores(org = NULL)

orgIndexFull()
}
\arguments{
\item{name}{org name}

\item{org}{org guid}

\item{.api}{default is 'twil'}

\item{.app}{TBD}

\item{extra_cols}{A logical indicating if new function columns should be returned}

\item{store}{store_short}
}
\description{
org functions
}
\section{Functions}{
\itemize{
\item \code{lookupOrgGuid()}: look up orgs guid given its' shortname. Returns NA if shortname not found

\item \code{orgAuth()}: look up orgs twil credentials given org Id

\item \code{orgScope()}: look up orgs twil credentials given org Id

\item \code{orgInfo()}: look up orgs twil credentials given org Id

\item \code{org_info()}: look up orgs twil credentials given org Id

\item \code{orgIntegrations()}: look up orgs twil credentials given org Id

\item \code{org_integrations()}: orgs

\item \code{orgShortName()}: orgs

\item \code{orgTimeZone()}: orgs

\item \code{get_store_id()}: orgs

\item \code{get_store_id_from_db()}: orgs

\item \code{get_org_stores()}: orgs

\item \code{orgIndexFull()}: org and store index

}}
\examples{
\dontrun{
lookupOrgGuid(name = "amuse")
}

}
