% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_systems.R
\name{cfg_systems}
\alias{cfg_systems}
\alias{erpSystemInfo}
\alias{lookup_erp_uuid}
\title{System functions}
\usage{
erpSystemInfo(erp)

lookup_erp_uuid(erp = NULL)
}
\arguments{
\item{erp}{erp system uuid}
}
\description{
System functions
}
\section{Functions}{
\itemize{
\item \code{erpSystemInfo()}: TBD

\item \code{lookup_erp_uuid()}: listing (index)

}}
