% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_users.R
\name{cfg_users}
\alias{cfg_users}
\alias{userOrg}
\alias{userScope}
\alias{userAuth}
\alias{userAccount}
\alias{userIntegrations}
\alias{userModsEnabled}
\title{Users}
\usage{
userOrg(usr, extra_cols = FALSE)

userScope(usr, .app = "polaris")

userAuth(usr, .api = "twil")

userAccount(usr)

userIntegrations(usr)

userModsEnabled(usr)
}
\arguments{
\item{usr}{user uuid}

\item{extra_cols}{A logical indicating if new function columns should be returned}

\item{.app}{TBD}

\item{.api}{TBD}
}
\description{
Users
}
\section{Functions}{
\itemize{
\item \code{userOrg()}: TBD

\item \code{userScope()}: TBD

\item \code{userAuth()}: TBD

\item \code{userAccount()}: TBD

\item \code{userIntegrations()}: TBD

\item \code{userModsEnabled()}: TBD

}}
