% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsconfig-package.R
\docType{package}
\name{rdsconfig-package}
\alias{rdsconfig}
\alias{rdsconfig-package}
\title{rdsconfig: Config Database Interface}
\description{
Functions that query the config database to retrieve data needed to administer and deploy hosted applications.
}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
