% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{..check_uuid}
\alias{sampleOrgs}
\alias{sampleUsers}
\alias{sampleLocation}
\alias{sampleLocationSku}
\title{Utilities}
\usage{
..check_uuid(id)

sampleOrgs(n = 5, resample = FALSE, ...)

sampleUsers(n = 5, resample = FALSE, ...)

sampleLocation()

sampleLocationSku()
}
\arguments{
\item{id}{uuid to check format}

\item{n}{sample size}

\item{resample}{replace each?}

\item{...}{additional params to function sample}
}
\description{
Utilities
}
\section{Functions}{
\itemize{
\item \code{..check_uuid()}: check format for valid uuid

\item \code{sampleOrgs()}: testing functions

\item \code{sampleUsers()}: testing functions

\item \code{sampleLocation()}: testing functions

\item \code{sampleLocationSku()}: testing functions

}}
